Refactors Watch app widget integration
Migrates to a new widget architecture for the Watch app, enhancing data handling
and widget updates.
This includes:
- Moving shared widget logic to a dedicated file.
- Updating data transfer objects for widget information.
- Simplifying data loading and refreshing in the main view model.
- Improving accessibility support for widgets.
- Corrects target for resources files.
