+++ MacMagazine/WatchApp/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
