+++ MacMagazine/WatchApp/WatchWidgetShared.swift
@@ -1,42 +1,42 @@
 import Foundation
 
 enum MacMagazineWidgetSharedStore {
+
     static let appGroupID = "group.com.brit.beta.macmagazine"
 
     private enum Keys {
-        static let lastPostTitle = "macmagazine.widget.lastPost.title"
-        static let lastPostDate = "macmagazine.widget.lastPost.date"
-        static let lastPostId = "macmagazine.widget.lastPost.id"
+        static let postId = "macmagazine.widget.post.id"
+        static let postTitle = "macmagazine.widget.post.title"
+        static let postDate = "macmagazine.widget.post.date"
     }
 
-    struct Snapshot: Equatable {
-        let postId: String
+    struct LastPost: Equatable {
+        let id: String
         let title: String
         let date: Date
     }
 
-    static func write(snapshot: Snapshot) {
+    static func write(post: LastPost) {
         guard let defaults = UserDefaults(suiteName: appGroupID) else { return }
 
-        defaults.set(snapshot.title, forKey: Keys.lastPostTitle)
-        defaults.set(snapshot.date, forKey: Keys.lastPostDate)
-        defaults.set(snapshot.postId, forKey: Keys.lastPostId)
-        defaults.synchronize()
+        defaults.set(post.id, forKey: Keys.postId)
+        defaults.set(post.title, forKey: Keys.postTitle)
+        defaults.set(post.date, forKey: Keys.postDate)
     }
 
-    static func readSnapshot() -> Snapshot? {
+    static func readPost() -> LastPost? {
         guard let defaults = UserDefaults(suiteName: appGroupID) else { return nil }
 
         guard
-            let title = defaults.string(forKey: Keys.lastPostTitle),
-            let date = defaults.object(forKey: Keys.lastPostDate) as? Date,
-            let postId = defaults.string(forKey: Keys.lastPostId),
-            !title.isEmpty,
-            !postId.isEmpty
+            let id = defaults.string(forKey: Keys.postId),
+            let title = defaults.string(forKey: Keys.postTitle),
+            let date = defaults.object(forKey: Keys.postDate) as? Date,
+            !id.isEmpty,
+            !title.isEmpty
         else {
             return nil
         }
 
-        return Snapshot(postId: postId, title: title, date: date)
+        return LastPost(id: id, title: title, date: date)
     }
 }
