+++ MacMagazine/WatchWidget/Extensions/DateFormatter.swift
@@ -1,0 +1,18 @@
+import Foundation
+
+extension DateFormatter {
+
+    static let watchWidgetTime: DateFormatter = {
+        let formatter = DateFormatter()
+        formatter.locale = Locale(identifier: "pt_BR")
+        formatter.dateFormat = "HH:mm"
+        return formatter
+    }()
+
+    static let watchWidgetDay: DateFormatter = {
+        let formatter = DateFormatter()
+        formatter.locale = Locale(identifier: "pt_BR")
+        formatter.dateFormat = "dd/MM"
+        return formatter
+    }()
+}
