+++ MacMagazine/WatchWidget/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
