+++ MacMagazine/WatchWidget/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
