+++ MacMagazine/WatchWidget/Resources/Assets.xcassets/Contents.json
