+++ MacMagazine/WatchWidget/Resources/Assets.xcassets/logo_color.imageset/Contents.json
