+++ MacMagazine/WatchWidget/Resources/Assets.xcassets/logo_color.imageset/logo@3x 3.png
