+++ MacMagazine/WatchWidget/Resources/Assets.xcassets/logo_white.imageset/Contents.json
