+++ MacMagazine/WatchWidget/Widget/AppIntent.swift
@@ -1,7 +1,7 @@
 import AppIntents
 import WidgetKit
 
-struct ConfigurationAppIntent: WidgetConfigurationIntent {
+struct AppIntent: WidgetConfigurationIntent {
     static var title: LocalizedStringResource { "MacMagazine" }
     static var description: IntentDescription { "Configuração da complicação." }
 }
